% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{tc_combine_edge_weights}
\alias{tc_combine_edge_weights}
\title{Combine the edge weights of two or more travelcost graphs}
\usage{
tc_combine_edge_weights(..., fun = "*")
}
\arguments{
\item{...}{: two or more \code{tc_graph} objects, or a list of such objects}

\item{fun}{function or string: the function (or function name) to be used to combine the edge weights. The function must operate on sparse matrices as returned by \code{\link[Matrix:sparseMatrix]{Matrix::sparseMatrix()}}}
}
\value{
A \code{tc_graph} object, which is the first object in \code{...} but with edge weights modified
}
\description{
Combine the edge weights of two or more travelcost graphs
}
\examples{
my_raster <- raster::raster(ext = raster::extent(c(152.75, 155.25, -56.25, -54.25)),
                            res = c(0.5, 0.5), crs = "+proj=longlat")
g <- tc_build_graph(my_raster)
## set random edge weights
g1 <- tc_set_edge_weights(g, values = runif(sum(tc_adj_matrix(g) > 0)))
g2 <- tc_set_edge_weights(g, values = runif(sum(tc_adj_matrix(g) > 0)))
## sum of those two
g3 <- tc_combine_edge_weights(g1, g2, fun = "+")

}
\seealso{
\code{\link[=tc_build_graph]{tc_build_graph()}}, \code{\link[=tc_combine_edge_weights]{tc_combine_edge_weights()}}
}
