% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost.R
\name{tc_cost}
\alias{tc_cost}
\title{Calculate the travel cost from one or more points to each point on a raster}
\usage{
tc_cost(x, from, direction = "out")
}
\arguments{
\item{x}{tc_graph: an object of class \code{tc_graph}, as returned by \code{\link[=tc_build_graph]{tc_build_graph()}}}

\item{from}{numeric or matrix: the points to calculate travel costs from. A two-element numeric vector or a 2-column matrix, or anything else accepted by \code{\link[raster:cellFrom]{raster::cellFromXY()}}}

\item{direction}{string:
\itemize{
\item "out" - calculate travel cost outwards from the 'from' points to each point on the raster
\item "in" - calculate travel cost from each point on the raster to each of the 'from' points
\item "both" the sum of "out" and "in"
}}
}
\value{
A raster layer (if \code{from} is one point) or brick (if \code{from} is more than one point)
}
\description{
Calculate the travel cost from one or more points to each point on a raster
}
\examples{
my_raster <- raster::raster(ext = raster::extent(c(153, 155, -56, -54)), res = c(0.5, 0.5),
                            crs = "+proj=longlat")
g <- tc_build_graph(my_raster)

## set random edge weights
g <- tc_set_edge_weights(g, values = runif(sum(tc_adj_matrix(g) > 0)))

## calculate travel cost
cx <- c(154, -55)
D <- tc_cost(g, from = cx)

}
\seealso{
\code{\link[=tc_build_graph]{tc_build_graph()}}
}
