% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d_movement.R
\name{tc_optimal_2d_velocity}
\alias{tc_optimal_2d_velocity}
\title{Calculate optimal 2D velocity}
\usage{
tc_optimal_2d_velocity(vw, phi, va, powerfun, refine = TRUE)
}
\arguments{
\item{vw}{scalar: the speed of the wind (or water, or other medium)}

\item{phi}{scalar: the angle (in radians) of the wind with respect to the desired direction of travel}

\item{va}{numeric: a vector of airspeeds to evaluate}

\item{powerfun}{function: a function that takes a single parameter (vector of airspeeds) and returns the power required for the animal to move at each of those speeds}

\item{refine}{logical: if \code{TRUE}, use the provided \code{va} values as a first guess and recursively call \code{\link[=tc_optimal_2d_velocity]{tc_optimal_2d_velocity()}} to refine the guess. If \code{FALSE}, return the optimal result from amongst the provided \code{va} values}
}
\value{
A data.frame with columns \code{va_opt}, \code{theta_opt} (the optimal airspeed and angle relative to the desired direction of travel), \code{vg} (the resultant ground speed in the desired direction of travel), and \code{E} the energy required (in J/m)
}
\description{
For movement such as flight or swimming, we often need to take account of winds or currents. These affect travel not only by speeding or slowing movement (e.g. tailwinds or headwinds), but also push the travel off course (winds blowing from the side). \code{\link[=tc_optimal_2d_velocity]{tc_optimal_2d_velocity()}} takes a function that estimates the power required by an animal to move at a given speed relative to its surrounding medium (air or water), along with the speed and direction of the wind or current, and estimates the most efficient speed (minimum energy per unit distance travelled) for the animal to travel at.
}
