% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travelcost-package.R
\docType{package}
\name{travelcost-package}
\alias{travelcost}
\alias{travelcost-package}
\title{travelcost: Estimate the cost of travel}
\description{
Estimate the cost of travel on a geographic grid, taking into account factors such as terrain, slope, or wind or current speeds.
}
\author{
\strong{Maintainer}: Ben Raymond \email{ben.raymond@aad.gov.au}

}
\keyword{internal}
